requirements.txt
# Core dependencies for protein science AI system
torch>=2.0.0
transformers>=4.30.0
biopython>=1.81
mdanalysis>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Structure prediction and analysis
fair-esm>=2.0.0
biotite>=0.39.0
pymol-open-source>=2.5.0

# Molecular simulation
openmm>=8.0.0
mdtraj>=1.9.0

# Database and web interfaces
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
gradio>=3.40.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Optional: GPU acceleration
# torch-geometric>=2.3.0
# torch-scatter>=2.1.0
# torch-sparse>=0.6.0

# Optional: Advanced structure prediction
# alphafold>=2.3.0
# colabfold>=1.5.0